# contracts/openapi/api.v1.yaml
openapi: 3.1.0
info:
  title: Industrial Knowledge Intelligence Platform pilot API
  version: 0.1.0-pilot
  description: >
    Runnable, provider-free decision-support prototype. It never controls equipment.
    Search results are ACL-filtered before ranking and answers are cited or abstain.
    Authentication is NOT production-ready: the required, caller-controlled X-Dev-Subject,
    X-Dev-Roles, and X-Dev-Sites headers are accepted only when IKIP_ENV=development and
    IKIP_DEV_AUTH=1. Values are never synthesized. No JWT verifier is included.
servers:
  - url: /
paths:
  /healthz:
    get:
      summary: Liveness check
      operationId: healthz
      responses:
        "200":
          description: Process is live.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties: { status: { type: string, const: ok } }
  /search:
    post:
      summary: Return authorized ranked evidence without answer synthesis.
      operationId: search
      security: [{ devSubject: [], devRoles: [], devSites: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/QueryRequest" } }
      responses:
        "200":
          description: Authorized evidence only.
          content:
            application/json: { schema: { $ref: "#/components/schemas/SearchResponse" } }
        "400": { description: A development identity header contains no usable value. }
        "403": { description: Identity is unverified; no content existence is disclosed. }
        "422": { description: A required development identity header or request field is missing. }
        "503": { description: Development auth is disabled and no production verifier exists. }
  /answer:
    post:
      summary: Return a templated, evidence-cited pilot answer or safe abstention.
      description: Uses DevAnswerGateway; no external or production model gateway is included.
      operationId: answer
      security: [{ devSubject: [], devRoles: [], devSites: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/QueryRequest" } }
      responses:
        "200":
          description: Grounded answer or abstention.
          content:
            application/json:
              schema: { $ref: "../schemas/answer.schema.json" }
        "400": { description: A development identity header contains no usable value. }
        "403": { description: Identity is unverified; no content existence is disclosed. }
        "422": { description: A required development identity header or request field is missing. }
        "503": { description: Development auth is disabled and no production verifier exists. }
  /documents:
    post:
      summary: Upload and process PDF, STL, or STEP multipart content.
      security: [{ devSubject: [], devRoles: [], devSites: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { type: object, required: [file], properties: { file: { type: string, format: binary }, sites: { type: string }, roles: { type: string } } }
      responses: { "202": { description: Original retained and processing scheduled. }, "413": { description: Too large. }, "422": { description: Invalid name/type/content. } }
    get:
      summary: List authorized uploaded documents and states.
      security: [{ devSubject: [], devRoles: [], devSites: [] }]
      responses: { "200": { description: Authorized documents only. } }
  /documents/{document_id}:
    get:
      summary: Get authorized processing status.
      security: [{ devSubject: [], devRoles: [], devSites: [] }]
      parameters: [{ in: path, name: document_id, required: true, schema: { type: string } }]
      responses: { "200": { description: Status. }, "404": { description: Absent or unauthorized. } }
  /documents/{document_id}/content:
    get:
      summary: Stream an authorized immutable original for preview.
      security: [{ devSubject: [], devRoles: [], devSites: [] }]
      parameters: [{ in: path, name: document_id, required: true, schema: { type: string } }]
      responses: { "200": { description: Original bytes. }, "404": { description: Absent or unauthorized. } }
  /admin/acl/revoke:
    post:
      summary: Revoke a seeded in-memory document ACL (development demonstration).
      description: Requires a verified development identity carrying the exact admin role.
      operationId: revokeAcl
      security: [{ devSubject: [], devRoles: [], devSites: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [document_id]
              properties: { document_id: { type: string, minLength: 1 } }
      responses:
        "200":
          description: Revocation was applied to this process's in-memory ACL store.
          content:
            application/json:
              schema:
                type: object
                required: [document_id, revoked]
                properties:
                  document_id: { type: string }
                  revoked: { type: boolean }
        "400": { description: A development identity header contains no usable value. }
        "403": { description: Verified admin role required. }
        "422": { description: A required development identity header or request field is missing. }
        "503": { description: Development auth is disabled and no production verifier exists. }
components:
  securitySchemes:
    devSubject:
      type: apiKey
      in: header
      name: X-Dev-Subject
      description: INSECURE, required development-only caller-controlled subject; never production auth.
    devRoles:
      type: apiKey
      in: header
      name: X-Dev-Roles
      description: INSECURE, required development-only comma-separated caller-controlled roles.
    devSites:
      type: apiKey
      in: header
      name: X-Dev-Sites
      description: INSECURE, required development-only comma-separated caller-controlled sites.
  schemas:
    QueryRequest:
      type: object
      required: [question]
      properties:
        question: { type: string, minLength: 1 }
        asset_ids: { type: array, items: { type: string }, default: [] }
        sites: { type: array, items: { type: string }, default: [] }
    Evidence:
      type: object
      required: [evidence_id, document_id, text, authority]
      properties:
        evidence_id: { type: string }
        document_id: { type: string }
        text: { type: string }
        authority: { type: string }
        retrieval_score: { type: [number, "null"] }
        retrieved_by: { type: [string, "null"] }
    SearchResponse:
      type: object
      required: [evidence, count]
      properties:
        evidence: { type: array, items: { $ref: "#/components/schemas/Evidence" } }
        count: { type: integer, minimum: 0 }